+++ src/lib.rs
@@ -1,4 +1,4 @@
-// Copyright (c) 2020 Weird Constructor <weirdconstructor@gmail.com>
+// Copyright (c) 2020-2022 Weird Constructor <weirdconstructor@gmail.com>
 // This is a part of WLambda. See README.md and COPYING for details.
 
 /*!
@@ -43,6 +43,21 @@ The embedding API and all internal operations rely on a data structure
 made of [VVal](https://docs.rs/wlambda/newest/wlambda/vval/index.html) nodes.
 
 Here you can find the [WLambda Language Reference](https://docs.rs/wlambda/newest/wlambda/prelude/index.html#wlambda-reference).
+
+# Compiling WLambda
+
+If you want to compile WLambda with all features enabled you need
+to run:
+
+```text
+    cargo build --features mqtt,http
+```
+
+or just:
+
+```text
+    cargo build --features all
+```
 
 # API Hello World
 
@@ -517,10 +532,10 @@ assert_eq!(r.s(), "42");
 Current remaining goals for WLambda are:
 
 - Fix remaining bugs.
-- Add missing standard library functions without dragging in more
+- DONE: Add missing standard library functions without dragging in more
 dependencies.
 - Improve and further document the VVal API for interacting with WLambda.
-- Improve [WLambda Language Reference](https://docs.rs/wlambda/newest/wlambda/prelude/index.html#wlambda-reference) documentation.
+- DONE: Improve [WLambda Language Reference](https://docs.rs/wlambda/newest/wlambda/prelude/index.html#wlambda-reference) documentation.
 - DONE: Complete function reference documentation in [WLambda Language Reference](https://docs.rs/wlambda/newest/wlambda/prelude/index.html#wlambda-reference).
 - DONE: Add proper module support (via `!@import` and `!@export`).
 - DONE: Add prototyped inheritance for OOP paradigm.