+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * **Feature:** (Optional) Embedded MQTT broker and client API for inter process
 communication across WLambda applications. See that you compile in the 
 `mqtt` feature and check out `std:mqtt:broker:new` and `std:mqtt:client:new`.
+* **Feature:** (Optional) HTTP Client API that supports GET for now. See
+`std:http:client:new` and `std:http:get`.
 * **Feature:** The LocalFileModuleResolver now does keep around the
 loaded symbol table and does not re-evaluate the module source on every
 import. This makes it possible to define global stuff inside the modules